requirements.txt
discord.py==2.4.0
pillow==10.3.0
requests
pymongo
datetime